refactor: extract MCP helper hot spots
