+++ crates/app/src/acp/acpx.rs
@@ -2,9 +2,9 @@ use std::collections::BTreeMap;
 use std::io::ErrorKind;
 use std::path::{Path, PathBuf};
 use std::process::Stdio;
-use std::sync::OnceLock;
 
 use async_trait::async_trait;
+#[cfg(test)]
 use base64::Engine;
 use serde::{Deserialize, Serialize};
 use serde_json::{Value, json};
@@ -14,9 +14,14 @@ use tokio::time::{Duration, Instant, sleep, sleep_until, timeout};
 
 use crate::CliResult;
 use crate::config::LoongClawConfig;
-use crate::mcp::{McpRegistry, McpStdioServerLaunchSpec};
 use crate::process_launch::resolve_command_invocation;
 
+#[cfg(test)]
+use super::acpx_mcp::{AcpxMcpServerEntry, AcpxMcpServerEnvEntry, build_mcp_proxy_agent_command};
+use super::acpx_mcp::{
+    build_mcp_proxy_agent_command_for_selection, injectable_mcp_server_count,
+    probe_mcp_proxy_support, validate_requested_mcp_servers,
+};
 use super::backend::{
     AcpAbortSignal, AcpBackendMetadata, AcpCapability, AcpConfigPatch, AcpDoctorReport,
     AcpRuntimeBackend, AcpSessionBootstrap, AcpSessionHandle, AcpSessionMode, AcpSessionState,
@@ -34,10 +39,6 @@ const ACPX_DEFAULT_QUEUE_OWNER_TTL_SECONDS: f64 = 0.1;
 const ACPX_PERMISSION_DENIED_EXIT_CODE: i32 = 5;
 const ACPX_SPAWN_RETRY_ATTEMPTS: usize = 5;
 const ACPX_SPAWN_RETRY_DELAY: Duration = Duration::from_millis(25);
-const ACPX_MCP_PROXY_NODE_COMMAND: &str = "node";
-const ACPX_MCP_PROXY_SCRIPT_NAME: &str = "loongclaw-acpx-mcp-proxy.mjs";
-const ACPX_MCP_PROXY_SCRIPT_SOURCE: &str = include_str!("assets/acpx-mcp-proxy.mjs");
-static ACPX_MCP_PROXY_SCRIPT_PATH: OnceLock<Result<String, String>> = OnceLock::new();
 
 #[derive(Debug, Clone, Serialize, Deserialize)]
 struct AcpxRuntimeHandleState {
@@ -77,22 +78,6 @@ struct AcpxIdentifiers {
     acpx_record_id: Option<String>,
     backend_session_id: Option<String>,
     agent_session_id: Option<String>,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-struct AcpxMcpServerEntry {
-    name: String,
-    command: String,
-    args: Vec<String>,
-    env: Vec<AcpxMcpServerEnvEntry>,
-    #[serde(default)]
-    cwd: Option<String>,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-struct AcpxMcpServerEnvEntry {
-    name: String,
-    value: String,
 }
 
 #[derive(Default)]
@@ -706,33 +691,6 @@ fn resolve_profile(config: &LoongClawConfig) -> CliResult<ResolvedAcpxProfile> {
     })
 }
 
-fn validate_requested_mcp_servers(
-    config: &LoongClawConfig,
-    request: &AcpSessionBootstrap,
-) -> CliResult<Vec<String>> {
-    if request.mcp_servers.is_empty() {
-        return Ok(Vec::new());
-    }
-    if !config.acp.allow_mcp_server_injection {
-        return Err(
-            "ACPX bootstrap requested MCP server injection but acp.allow_mcp_server_injection=false"
-                .to_owned(),
-        );
-    }
-
-    let registry = McpRegistry::from_config(config)?;
-    let selected = registry.resolve_selected_server_names(&request.mcp_servers)?;
-    let _launch_specs = registry.resolve_injectable_stdio_launch_specs(&selected)?;
-
-    Ok(selected)
-}
-
-fn injectable_mcp_server_count(config: &LoongClawConfig) -> CliResult<usize> {
-    let registry = McpRegistry::from_config(config)?;
-    let count = registry.injectable_stdio_server_count();
-    Ok(count)
-}
-
 async fn build_verb_args<I>(
     config: &LoongClawConfig,
     profile: &ResolvedAcpxProfile,
@@ -812,8 +770,11 @@ async fn resolve_raw_agent_command(
     }
 
     let target_command = resolve_acpx_agent_command(profile, timeout_ms, cwd, agent).await?;
-    let mcp_servers = resolve_selected_mcp_server_entries(config, selected_mcp_servers)?;
-    let proxy_command = build_mcp_proxy_agent_command(target_command.as_str(), &mcp_servers)?;
+    let proxy_command = build_mcp_proxy_agent_command_for_selection(
+        config,
+        target_command.as_str(),
+        selected_mcp_servers,
+    )?;
     Ok(Some(proxy_command))
 }
 
@@ -882,148 +843,6 @@ fn builtin_agent_command(agent: &str) -> Option<String> {
         _ => return None,
     };
     Some(command.to_owned())
-}
-
-fn resolve_selected_mcp_server_entries(
-    config: &LoongClawConfig,
-    selected_mcp_servers: &[String],
-) -> CliResult<Vec<AcpxMcpServerEntry>> {
-    let registry = McpRegistry::from_config(config)?;
-    let resolved_servers = registry.resolve_injectable_stdio_launch_specs(selected_mcp_servers)?;
-
-    let mut entries = Vec::new();
-    for server in resolved_servers {
-        let entry = acpx_mcp_server_entry_from_registry_server(server);
-        entries.push(entry);
-    }
-
-    Ok(entries)
-}
-
-fn acpx_mcp_server_entry_from_registry_server(
-    server: McpStdioServerLaunchSpec,
-) -> AcpxMcpServerEntry {
-    let mut env_entries = Vec::new();
-    for (name, value) in server.env {
-        let env_entry = AcpxMcpServerEnvEntry { name, value };
-        env_entries.push(env_entry);
-    }
-
-    AcpxMcpServerEntry {
-        name: server.name,
-        command: server.command,
-        args: server.args,
-        env: env_entries,
-        cwd: server.cwd,
-    }
-}
-
-fn build_mcp_proxy_agent_command(
-    target_command: &str,
-    mcp_servers: &[AcpxMcpServerEntry],
-) -> CliResult<String> {
-    let script_path = ensure_mcp_proxy_script_path()?;
-    let payload = serde_json::to_vec(&json!({
-        "targetCommand": target_command,
-        "mcpServers": mcp_servers,
-    }))
-    .map_err(|error| format!("serialize ACPX MCP proxy payload failed: {error}"))?;
-    let encoded = base64::engine::general_purpose::URL_SAFE_NO_PAD.encode(payload);
-    Ok(join_command_line(&[
-        ACPX_MCP_PROXY_NODE_COMMAND.to_owned(),
-        script_path,
-        "--payload".to_owned(),
-        encoded,
-    ]))
-}
-
-fn ensure_mcp_proxy_script_path() -> CliResult<String> {
-    ACPX_MCP_PROXY_SCRIPT_PATH
-        .get_or_init(materialize_mcp_proxy_script)
-        .clone()
-}
-
-fn materialize_mcp_proxy_script() -> Result<String, String> {
-    let path = std::env::temp_dir()
-        .join("loongclaw")
-        .join(ACPX_MCP_PROXY_SCRIPT_NAME);
-    if let Some(parent) = path.parent() {
-        std::fs::create_dir_all(parent)
-            .map_err(|error| format!("create ACPX MCP proxy directory failed: {error}"))?;
-    }
-    std::fs::write(&path, ACPX_MCP_PROXY_SCRIPT_SOURCE)
-        .map_err(|error| format!("write ACPX MCP proxy script failed: {error}"))?;
-    #[cfg(unix)]
-    {
-        use std::os::unix::fs::PermissionsExt;
-
-        let mut permissions = std::fs::metadata(&path)
-            .map_err(|error| format!("stat ACPX MCP proxy script failed: {error}"))?
-            .permissions();
-        permissions.set_mode(0o755);
-        std::fs::set_permissions(&path, permissions)
-            .map_err(|error| format!("chmod ACPX MCP proxy script failed: {error}"))?;
-    }
-    Ok(path.display().to_string())
-}
-
-async fn probe_mcp_proxy_support(cwd: Option<&str>) -> CliResult<(String, String)> {
-    let script_path = ensure_mcp_proxy_script_path()?;
-    let mut probe = Command::new(ACPX_MCP_PROXY_NODE_COMMAND);
-    probe.arg("--version");
-    if let Some(cwd) = cwd {
-        probe.current_dir(cwd);
-    }
-    let output = probe.output().await.map_err(|error| {
-        if error.kind() == ErrorKind::NotFound {
-            format!("embedded ACPX MCP proxy requires `{ACPX_MCP_PROXY_NODE_COMMAND}` on PATH")
-        } else {
-            format!("probe embedded ACPX MCP proxy runtime failed: {error}")
-        }
-    })?;
-    let stdout = String::from_utf8_lossy(&output.stdout).trim().to_owned();
-    let stderr = String::from_utf8_lossy(&output.stderr).trim().to_owned();
-    let observed = match (stdout.is_empty(), stderr.is_empty()) {
-        (false, true) => stdout,
-        (true, false) => stderr,
-        (false, false) => format!("{stdout} | {stderr}"),
-        (true, true) => "(empty)".to_owned(),
-    };
-    if !output.status.success() {
-        return Err(format!(
-            "embedded ACPX MCP proxy runtime probe exited with code {}: {observed}",
-            output
-                .status
-                .code()
-                .map_or_else(|| "unknown".to_owned(), |code| code.to_string())
-        ));
-    }
-    Ok((script_path, observed))
-}
-
-fn join_command_line(parts: &[String]) -> String {
-    parts
-        .iter()
-        .map(|part| quote_command_part(part.as_str()))
-        .collect::<Vec<_>>()
-        .join(" ")
-}
-
-fn quote_command_part(value: &str) -> String {
-    if value.is_empty() {
-        return "\"\"".to_owned();
-    }
-    if value.chars().all(|ch| {
-        ch.is_ascii_alphanumeric()
-            || matches!(
-                ch,
-                '_' | '.' | '/' | ':' | '@' | '%' | '+' | '=' | ',' | '-'
-            )
-    }) {
-        return value.to_owned();
-    }
-    let escaped = value.replace('\\', "\\\\").replace('"', "\\\"");
-    format!("\"{escaped}\"")
 }
 
 fn resolve_effective_cwd(
